Move validate_schema out of module root, add tests
